# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TUSS API",
    "description": "\nConsulta às tabelas TUSS da ANS com histórico de versões e vigência por data.\n\nToda rota `/v1` exige `Authorization: Bearer <token>`. Erros seguem o formato\nProblem Details (RFC 9457) e trazem um `request_id` para rastrear no log.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/status": {
      "get": {
        "summary": "Última sincronização de cada tabela",
        "description": "Mostra, por tabela, a carga que a API está servindo e a tentativa de carga mais recente.\n\nÉ a forma de saber se os dados estão atualizados e se alguma carga falhou.",
        "operationId": "status_v1_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/tabelas": {
      "get": {
        "summary": "Tabelas carregadas",
        "operationId": "listar_tabelas_v1_tabelas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaTabelas"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/tabelas/{tabela}/conceitos": {
      "get": {
        "summary": "Conceitos de uma tabela, em ordem de código",
        "description": "Sem `q`: todos, paginados por cursor (repita a chamada com `cursor=proximo_cursor`).\n\nCom `q`: busca, dos mais relevantes para os menos, até `limite` resultados.",
        "operationId": "listar_conceitos_v1_tabelas__tabela__conceitos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tabela",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 20,
              "description": "`tuss-22` ou só `22`",
              "examples": [
                "22"
              ],
              "title": "Tabela"
            },
            "description": "`tuss-22` ou só `22`"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 3,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Busca. Código ou começo de código (`1010`, `1.01.01.01-2`) ou texto da descrição, sem precisar de acento e tolerando erro de digitação. Devolve os mais relevantes primeiro, até `limite`, sem próxima página.",
              "examples": [
                "consulta consultorio"
              ],
              "title": "Q"
            },
            "description": "Busca. Código ou começo de código (`1010`, `1.01.01.01-2`) ou texto da descrição, sem precisar de acento e tolerando erro de digitação. Devolve os mais relevantes primeiro, até `limite`, sem próxima página."
          },
          {
            "name": "vigente_em",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Só os conceitos vigentes nesta data (AAAA-MM-DD), cada um como era nela. Não se combina com `q`.",
              "title": "Vigente Em"
            },
            "description": "Só os conceitos vigentes nesta data (AAAA-MM-DD), cada um como era nela. Não se combina com `q`."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Valor de `proximo_cursor` da página anterior",
              "title": "Cursor"
            },
            "description": "Valor de `proximo_cursor` da página anterior"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Itens por página",
              "default": 50,
              "title": "Limite"
            },
            "description": "Itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaConceitos"
                }
              }
            }
          },
          "404": {
            "description": "Tabela ou código não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tabelas/{tabela}/conceitos/{codigo}": {
      "get": {
        "summary": "Um conceito pelo código (estado atual ou numa data)",
        "description": "Sem `em`: a versão atual, e se está vigente hoje. Com `em`: a versão que valia na data.\n\nVigência usa as datas da ANS (`criterio: oficial`); sem elas, o período em que o\ncódigo apareceu nas cargas (`criterio: observado`). Antes da primeira carga, a\nversão mais antiga que conhecemos.",
        "operationId": "consultar_conceito_v1_tabelas__tabela__conceitos__codigo__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tabela",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 20,
              "description": "`tuss-22` ou só `22`",
              "examples": [
                "22"
              ],
              "title": "Tabela"
            },
            "description": "`tuss-22` ou só `22`"
          },
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "examples": [
                "10101012"
              ],
              "title": "Codigo"
            }
          },
          {
            "name": "em",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data (AAAA-MM-DD): devolve o código como era nesse dia e se estava vigente. 404 se nesse dia o código estava fora da lista.",
              "title": "Em"
            },
            "description": "Data (AAAA-MM-DD): devolve o código como era nesse dia e se estava vigente. 404 se nesse dia o código estava fora da lista."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceitoDetalhe"
                }
              }
            }
          },
          "404": {
            "description": "Tabela ou código não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tabelas/{tabela}/conceitos/{codigo}/historico": {
      "get": {
        "summary": "Todas as versões de um código",
        "description": "Cada estado que o código já teve na nossa base, com o período em que valeu.\n\nInclui códigos que saíram da lista (a última versão tem `publicado_ate`).",
        "operationId": "historico_v1_tabelas__tabela__conceitos__codigo__historico_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tabela",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 20,
              "description": "`tuss-22` ou só `22`",
              "examples": [
                "22"
              ],
              "title": "Tabela"
            },
            "description": "`tuss-22` ou só `22`"
          },
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "examples": [
                "10101012"
              ],
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Historico"
                }
              }
            }
          },
          "404": {
            "description": "Tabela ou código não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validacoes": {
      "post": {
        "summary": "Confere se códigos estavam vigentes em datas",
        "description": "Até 100 itens de tabela + código + data; para cada um, se estava vigente e por quê.\n\nUm item com problema (tabela ou código que não existe) não derruba os outros: vem\ncom `vigente: false` e o `motivo`. Conta no limite de buscas por minuto.",
        "operationId": "validar_v1_validacoes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoValidacao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoValidacoes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/mudancas": {
      "get": {
        "summary": "Feed de mudanças publicadas",
        "description": "Inclusões, alterações, remoções e reativações, da mais antiga para a mais nova.\n\nPara acompanhar: guarde o `proximo_cursor` da última página e volte com ele depois.\nA carga inicial de cada tabela não gera mudanças (é o ponto de partida do histórico).",
        "operationId": "mudancas_v1_mudancas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A partir deste momento (AAAA-MM-DD ou data e hora; sem fuso = UTC)",
              "examples": [
                "2026-09-01"
              ],
              "title": "Desde"
            },
            "description": "A partir deste momento (AAAA-MM-DD ou data e hora; sem fuso = UTC)"
          },
          {
            "name": "tabela",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "description": "`tuss-22` ou só `22`",
              "title": "Tabela"
            },
            "description": "`tuss-22` ou só `22`"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "incluido",
                    "alterado",
                    "removido",
                    "reativado"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Valor de `proximo_cursor` da página anterior",
              "title": "Cursor"
            },
            "description": "Valor de `proximo_cursor` da página anterior"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Itens por página",
              "default": 50,
              "title": "Limite"
            },
            "description": "Itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaMudancas"
                }
              }
            }
          },
          "404": {
            "description": "Tabela ou código não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConceitoDetalhe": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo",
            "examples": [
              "10101012"
            ]
          },
          "descricao": {
            "type": "string",
            "title": "Descricao",
            "examples": [
              "Consulta em consultório (no horário normal ou preestabelecido)"
            ]
          },
          "inicio_vigencia": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inicio Vigencia"
          },
          "fim_vigencia": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fim Vigencia",
            "description": "Vazio = período aberto"
          },
          "fim_implantacao": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fim Implantacao"
          },
          "atributos": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Atributos",
            "description": "Campos próprios de cada tabela (ex.: laboratorio na tuss-20)"
          },
          "criterio": {
            "type": "string",
            "enum": [
              "oficial",
              "observado"
            ],
            "title": "Criterio",
            "description": "oficial: datas informadas pela ANS; observado: período visto nas cargas"
          },
          "tabela": {
            "type": "string",
            "title": "Tabela",
            "examples": [
              "tuss-22"
            ]
          },
          "carga_id": {
            "type": "integer",
            "title": "Carga Id",
            "description": "Carga publicada de onde veio a resposta"
          },
          "sincronizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sincronizado Em",
            "description": "Última vez que a fonte foi conferida"
          },
          "em": {
            "type": "string",
            "format": "date",
            "title": "Em",
            "description": "Data consultada: a de `?em=` ou, sem ela, hoje (UTC)"
          },
          "vigente": {
            "type": "boolean",
            "title": "Vigente",
            "description": "Se o código estava vigente na data consultada"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "descricao",
          "inicio_vigencia",
          "fim_vigencia",
          "fim_implantacao",
          "atributos",
          "criterio",
          "tabela",
          "carga_id",
          "sincronizado_em",
          "em",
          "vigente"
        ],
        "title": "ConceitoDetalhe"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListaTabelas": {
        "properties": {
          "itens": {
            "items": {
              "$ref": "#/components/schemas/Tabela"
            },
            "type": "array",
            "title": "Itens"
          }
        },
        "type": "object",
        "required": [
          "itens"
        ],
        "title": "ListaTabelas"
      },
...
      "PaginaConceitos": {
        "properties": {
          "tabela": {
            "type": "string",
            "title": "Tabela"
          },
          "carga_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carga Id"
          },
          "sincronizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sincronizado Em"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/Conceito"
            },
            "type": "array",
            "title": "Itens"
          },
          "proximo_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proximo Cursor",
            "description": "Passe em `cursor` para a próxima página; vazio = acabou"
          }
        },
        "type": "object",
        "required": [
          "tabela",
          "carga_id",
          "sincronizado_em",
          "itens",
          "proximo_cursor"
        ],
        "title": "PaginaConceitos"
      },
...
      "Status": {
        "properties": {
          "tabelas": {
            "items": {
              "$ref": "#/components/schemas/StatusTabela"
            },
            "type": "array",
            "title": "Tabelas"
          }
        },
        "type": "object",
        "required": [
          "tabelas"
        ],
        "title": "Status"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token gerado com `tuss token`",
        "scheme": "bearer"
      }
    }
  }
}